Extracted assembly helpers (#94)
* Moved methods used for handler registration in shared static helper
* Added missing extension methods for handler registration
