+++ src/Workleap.DomainEventPropagation.Subscription/Workleap.DomainEventPropagation.Subscription.csproj
@@ -46,5 +46,6 @@
     </Compile>
     <Compile Include="..\Shared\DomainEventWrapper.cs" Link="Shared\DomainEventWrapper.cs" />
     <Compile Include="..\Shared\TracingHelper.cs" Link="Shared\TracingHelper.cs" />
+    <Compile Include="..\Shared\AssemblyHelper.cs" Link="Shared\AssemblyHelper.cs" />
   </ItemGroup>
 </Project>
